split out config_csproj,  work on building out embedded csproj for ios/android
