+++ src/gen_build/gen_build/CustomBuild.cs
@@ -6,25 +6,28 @@ using System.Threading.Tasks;
 
 namespace GenBuild
 {
-	// TODO Create .snk for build if it does not exists
-	// TODO Don't keep generating items 
-	// TODO Generate tests
+	// TODO Create .snk for build if it does not exists - right now just use 'sn -k ...name...'
 	// TODO Generate bundle from dylib/so/etc.
+	// TODO Generate tests?
 	public class CustomBuild
 	{
 		private string _name;
-		private string _simple;
-		private string _actual;
+		private string _libRoot;
 
-		public CustomBuild(string name) : this(name, name, name)
-		{
+		public CustomBuild(string name)
+			: this(name, null)
+		{		
 		}
 
-		public CustomBuild(string name, string simple, string actual)
+		/// <summary>
+		/// 
+		/// </summary>
+		/// <param name="name"></param>
+		/// <param name="libRoot">e.g., C:\Users\tim\ownCloud\Loqu8\ext - containing sqlite\macosx\universal\lib\libxsqlite3.dylib</param>
+		public CustomBuild(string name, string libRoot)
 		{
 			_name = name;
-			_simple = name;
-			_actual = name;
+			_libRoot = libRoot;
 		}
 
 		private List<config_sqlite3> _itemsSqlite3;
@@ -87,6 +90,11 @@ namespace GenBuild
 
 			//items_sqlite3.Add(new config_sqlite3 { toolset = "v120_wp81", cpu = "arm" });
 			//items_sqlite3.Add(new config_sqlite3 { toolset = "v120_wp81", cpu = "x86" });
+
+			foreach (config_sqlite3 cfg in items_sqlite3)
+			{
+				cfg.guid = "{" + Guid.NewGuid().ToString().ToUpper() + "}";
+			}
 			return items_sqlite3;
 		}
 
@@ -95,6 +103,11 @@ namespace GenBuild
 			var items_cppinterop = new List<config_cppinterop>();
 			//items_cppinterop.Add(new config_cppinterop { env = "wp80", cpu = "arm" });
 			//items_cppinterop.Add(new config_cppinterop { env = "wp80", cpu = "x86" });
+
+			foreach (config_cppinterop cfg in items_cppinterop)
+			{
+				cfg.guid = "{" + Guid.NewGuid().ToString().ToUpper() + "}";
+			}
 			return items_cppinterop;
 		}
 
@@ -103,14 +116,15 @@ namespace GenBuild
 			var items_csproj = new List<config_csproj>();
 			genCoreItemsCsproj(items_csproj);
 
-			items_csproj.Add(config_csproj.create_provider(name, "netstandard11"));
-			items_csproj.Add(config_csproj.create_provider(name, "net35"));
-			items_csproj.Add(config_csproj.create_provider(name, "net40"));
+			// generate provider
+			//items_csproj.Add(config_csproj.create_provider(name, "netstandard11"));
+			//items_csproj.Add(config_csproj.create_provider(name, "net35"));
+			//items_csproj.Add(config_csproj.create_provider(name, "net40"));
 			items_csproj.Add(config_csproj.create_provider(name, "net45"));
 			items_csproj.Add(config_csproj.create_provider(name, "android"));
-			items_csproj.Add(config_csproj.create_provider(name, "win8"));
-			items_csproj.Add(config_csproj.create_provider(name, "win81"));
-			items_csproj.Add(config_csproj.create_provider(name, "wpa81"));
+			//items_csproj.Add(config_csproj.create_provider(name, "win8"));
+			//items_csproj.Add(config_csproj.create_provider(name, "win81"));
+			//items_csproj.Add(config_csproj.create_provider(name, "wpa81"));
 			items_csproj.Add(config_csproj.create_provider(name, "uwp10"));
 			items_csproj.Add(config_csproj.create_provider(name, "macos"));
 			// ios would only make sense here with dylibs - prefer internal
@@ -118,6 +132,31 @@ namespace GenBuild
 
 			items_csproj.Add(config_csproj.create_provider("internal", "ios_unified"));
 
+			if (!string.IsNullOrEmpty(_libRoot)) {
+				// generate batteries
+				items_csproj.Add(config_csproj.create_embedded(_name, "android"));
+				items_csproj.Add(config_csproj.create_embedded(_name, "ios_unified"));
+
+				// e.g., bundle_xsqlite3
+				var ver = 2;
+				var area = string.Format("batteries_{0}", _name);
+				items_csproj.Add(config_csproj.create_internal_batteries(area, ver, "ios_unified", _name));
+				// TODO items_csproj.Add(config_csproj.create_internal_batteries("batteries_e_sqlite3", "watchos", "e_sqlite3"));
+				items_csproj.Add(config_csproj.create_batteries(area, ver, "android", _name));
+				items_csproj.Add(config_csproj.create_batteries(area, ver, "macos", _name));
+				//items_csproj.Add(config_csproj.create_batteries(area, ver, "win8", _name));
+				//items_csproj.Add(config_csproj.create_batteries(area, ver, "wpa81", _name));
+				//items_csproj.Add(config_csproj.create_batteries(area, ver, "win81", _name));
+				items_csproj.Add(config_csproj.create_batteries(area, ver, "uwp10", _name));
+				//items_csproj.Add(config_csproj.create_batteries(area, ver, "net35", _name));
+				//items_csproj.Add(config_csproj.create_batteries(area, ver, "net40", _name));
+				items_csproj.Add(config_csproj.create_batteries(area, ver, "net45", _name));
+			}
+
+			foreach (config_csproj cfg in items_csproj)
+			{
+				cfg.guid = "{" + Guid.NewGuid().ToString().ToUpper() + "}";
+			}
 			return items_csproj;
 		}
